+++ src/lib/components/table/proposals/ResolvedHeight.tsx
@@ -6,18 +6,15 @@ import type { ProposalsTableRowProps } from "./ProposalsTableRow";
 
 export const ResolvedHeight = ({
   resolvedHeight,
-  isDepositFailed,
   isDepositOrVoting,
   amptrackSection,
 }: {
   resolvedHeight: ProposalsTableRowProps["proposal"]["resolvedHeight"];
-  isDepositFailed: boolean;
   isDepositOrVoting: boolean;
   amptrackSection?: string;
 }) => {
   if (isDepositOrVoting) return <Text color="text.dark">Pending</Text>;
-  if (!resolvedHeight || isDepositFailed)
-    return <Text color="text.dark">N/A</Text>;
+  if (!resolvedHeight) return <Text color="text.dark">N/A</Text>;
   return (
     <ExplorerLink
       type="block_height"